Refactoring to use BugDecimal for payment instead Double

+++ ecomarkets/src/main/java/ecomarkets/domain/core/basket/BasketItem.java
@@ -1,5 +1,6 @@
 package ecomarkets.domain.core.basket;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import ecomarkets.domain.core.product.ProductId;
@@ -8,10 +9,10 @@ import jakarta.persistence.Embeddable;
 @Embeddable
 public record BasketItem (ProductId productId,
         Integer amount,
-        Double totalPayment,
+        BigDecimal totalPayment,
         LocalDateTime creationDate){
 
-    public static BasketItem of(ProductId productId, Integer amount, Double totalPayment){
+    public static BasketItem of(ProductId productId, Integer amount, BigDecimal totalPayment){
         BasketItem bi = new BasketItem(productId, amount, totalPayment, LocalDateTime.now());
         return bi;
     }
