+++ ecomarkets/src/main/java/ecomarkets/domain/notification/email/EmailTemplate.java
@@ -1,5 +1,7 @@
 package ecomarkets.domain.notification.email;
 
+import java.math.BigDecimal;
+
 public interface EmailTemplate {
-    String getBody(String partnerName, Long basketId, Double paymentValue, String status);
+    String getBody(String partnerName, Long basketId, BigDecimal paymentValue, String status);
 }
